# Pixel as centroids optimization