… raster/test/regress/rt_pixelascentroids_expected