- `ST_PixelAsCentroids`, `ST_PixelAsCentroid` rewritten to use new C function  `RASTER_getPixelCentroids` (mostly copied from `RASTER_getPixelPolygons`).
- Test updated to add a case for rotated raster.

`ST_PixelAsCentroids(rast, 1)` execution time comparison for some sizes of raster:
| Raster size | Exec. time before | Exec. time after |
|-------------|-------------------|------------------|
| 100x100 | 44.677 ms | 4.228 ms |
| 200x200 | 183.234 ms | 17.441 ms |
| 500x500 | 1144.930 ms | 141.880 ms
| 1000x1000 | 4630.656 ms | 573.416 ms
| 2000x2000 | 18302.578 ms | 2290.872 ms
| 5000x5000 | 115014.318 ms | 14121.408 ms

(On Intel Core i7-11370H @ 3.30GHz × 8, 16 Gb RAM)